refactor: improve exception handling in team routes

+++ src/controller/teams-controllers.ts
@@ -1,6 +1,7 @@
 import { Request, Response, NextFunction } from "express";
 import { prisma } from "@/database/prisma";
 import { z } from "zod"
+import { AppError } from "@/utils/App.Error";
 
 class TeamsController {
   async create(requeste:Request, response:Response, next:NextFunction) {
@@ -11,19 +12,26 @@ class TeamsController {
 
     const { name, description} = bodySchema.parse(requeste.body)
 
+    const varifyTeamName = await prisma.team.findFirst({where: {name}})
+
+    if(varifyTeamName){
+      throw new AppError("Nome do time não disponivel",400)
+    }
+
     const team = await prisma.team.create({
       data: {
         name,
         description
       }
     })
-    return response.json({team})
+    return response.status(201).json({team})
   }
 
   async show(request:Request, response:Response, next:NextFunction) {
+
     const teams =  await prisma.team.findMany({select: {name: true, id: true, description: true}})
 
-    return response.json(teams)
+    return response.status(200).json(teams)
   }
 
   async update(request:Request, response:Response, next:NextFunction) {
@@ -36,8 +44,14 @@ class TeamsController {
       description: z.string().optional()
     })
 
-
     const { id } = paramSchema.parse(request.params)
+
+    const verifyIdExist = await prisma.team.findUnique({where: {id}})
+
+    if(!verifyIdExist){
+      throw new AppError("ID NotFound", 404)
+    }
+
     const { name, description } = bodySchema.parse(request.body)
 
     const teamUpdate = await prisma.team.update({
@@ -50,7 +64,7 @@ class TeamsController {
       }
     })
     
-    return response.json(teamUpdate)
+    return response.status(200).json(teamUpdate)
   }
 
   async delete(request:Request, response:Response, next:NextFunction) {
@@ -60,10 +74,16 @@ class TeamsController {
 
     const { id } = paramSchema.parse(request.params)
 
+    const verifyIdExist = await prisma.team.findUnique({where: {id}})
+
+    if(!verifyIdExist){
+      throw new AppError("ID Notfound",404)
+    }
+
     const deleteTeam = await prisma.team.delete({
       where: {id}
     })
-    return response.status(201).json({message: "sucess"})
+    return response.status(204).json({message: "sucess"})
   }
 }
